+++ 3-STUDENTS/sotnichenko_ekaterina/project/src/components/index.js
@@ -1,0 +1,7 @@
+import Basket from './basket.js'
+import Catalog from './catalog.js'
+
+export default () => {
+    let basket = new Basket();
+    let catalog = new Catalog(basket);
+}
