+++ 3-STUDENTS/sotnichenko_ekaterina/project/src/index.js
@@ -1,0 +1,6 @@
+import app from './components'
+import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import './layout/styles/style.css';
+
+app();
